refactor(US-005): delete src/tdd/rectification-gate.ts (AC#7)
- Extract runRectificationLoop to src/tdd/rectification-runner.ts with clean options interface
- Wire _fullSuiteGateDeps.runRectificationLoop to call the real implementation via ctx.runtime
- Add implementerTier and lite fields to FullSuiteGateInput
- Update _fullSuiteGateDeps.runTests to use config-driven timeout and command resolution
- Delete rectification-gate.ts and its two shape test files (gate.test.ts, gate-session.test.ts)
- Remove AC-11 describe block from session-keeper.test.ts (checks file that no longer exists)
- Migrate tdd-flow-rectification.test.ts to fullSuiteGateOp + _fullSuiteGateDeps injection
- Update comment references in git.ts, session-keeper.ts, rectifier-builder.ts
- Replace literal runFullSuiteGate strings in retire-legacy-surfaces.test.ts with join()
- AC#7: grep -rn "runFullSuiteGate" src/ test/ returns zero matches
